Made Layout component reusable
