+++ client/src/components/admin/Admin.js
@@ -1,23 +1,35 @@
 import React, { Component } from "react";
-import Layout from "./layout/Layout";
+import Layout from "../layout/Layout";
 import { Route } from "react-router-dom";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import Properties from "./properties/Properties";
 import WorkOrders from "./workorders/WorkOrders";
-import Settings from "./settings/Settings";
+import SettingsPage from "./settings/SettingsPage";
 import Billing from "./billing/Billing";
 import NewTenant from "./tenants/NewTenants";
+import Business from "@material-ui/icons/Business";
+import Assignment from "@material-ui/icons/Assignment";
+import People from "@material-ui/icons/People";
+import Settings from "@material-ui/icons/Settings";
+import CreditCard from "@material-ui/icons/CreditCard";
+
+const links = [
+  { name: "Properties", url: "properties", icon: <Business /> },
+  { name: "Work Orders", url: "work-orders", icon: <Assignment /> },
+  { name: "Tenants", url: "tenants", icon: <People /> },
+  { name: "Billing", url: "billing", icon: <CreditCard /> },
+  { name: "Settings", url: "settings", icon: <Settings /> }
+];
 
 class Admin extends Component {
-  state = {};
   render() {
     return (
-      <Layout>
+      <Layout links={links}>
         <CssBaseline />
         <Route exact path="/admin/properties" component={Properties} />
         <Route exact path="/admin/work-orders" component={WorkOrders} />
         <Route exact path="/admin/tenants" component={NewTenant} />
-        <Route exact path="/admin/settings" component={Settings} />
+        <Route exact path="/admin/settings" component={SettingsPage} />
         <Route exact path="/admin/billing" component={Billing} />
       </Layout>
     );
