+++ client/src/components/layout/Layout.js
@@ -2,12 +2,12 @@ import React, { Component } from "react";
 import SidebarAdmin from "./SidebarAdmin";
 import Grid from "@material-ui/core/Grid";
 import Footer from "./Footer";
+
 class Layout extends Component {
-  state = {};
   render() {
     return (
       <Grid container justify="center" spacing={16}>
-        <SidebarAdmin />
+        <SidebarAdmin links={this.props.links} />
 
         <Grid item xs={10}>
           {this.props.children}
